# Conflicts:
#	WalletKitCore/src/bitcoin/BRSyncManager.c
#	WalletKitCore/src/generic/BRGenericManager.c
#	WalletKitJava/corecrypto/src/main/java/com/breadwallet/corecrypto/System.java
#	WalletKitSwift/WalletKit/BRCryptoSystem.swift

+++ WalletKitCore/src/ethereum/ewm/BREthereumEWMPrivate.h
@@ -453,7 +453,7 @@ ewmSignalAnnounceTransaction(BREthereumEWM ewm,
 typedef struct {
     BREthereumHash hash;
     BREthereumAddress contract;
-    int topicCount;
+    size_t topicCount;
     char **arrayTopics;
     char *data;
     UInt256 gasPrice;
